refactor: OrderApplyCount와 OrderResult 패키지 분리
- 패키지 분리 및 데이터소스 분리
