+++ order/src/main/java/com/watermelon/server/common/dataSourceConfig/BasicDatasourceConfig.java
@@ -1,31 +1,32 @@
-package com.watermelon.server.common.config;
-
+package com.watermelon.server.common.dataSourceConfig;
 
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.orm.jpa.EntityManagerFactoryBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
+import javax.sql.DataSource;
 import java.util.HashMap;
 import java.util.Map;
 
 @EnableTransactionManagement
 @EnableJpaRepositories(
-        entityManagerFactoryRef = "orderResultEntityManager",
-        transactionManagerRef = "orderResultTransactionManager",
-        basePackages = {"com.watermelon.server.orderResult.repository"}
+        entityManagerFactoryRef = "entityManagerFactory",
+        transactionManagerRef = "transactionManager",
+        basePackages = {"com.watermelon.server.order.repository"}
 )
 @Configuration
 @Profile("!local")
-public class OrderResultDataSourceConfig {
+public class BasicDatasourceConfig {
+
     @Value("${spring.datasource.url}")
     private String url;
 
@@ -35,43 +36,51 @@ public class OrderResultDataSourceConfig {
     @Value("${spring.datasource.password}")
     private String password;
 
+    @Value("${spring.datasource.hikari.maximum-pool-size}")
+    private int maximumPoolSize;
+
+    @Value("${spring.datasource.hikari.leak-detection-threshold}")
+    private long leakDetectionThreshold;
+
     @Value("${spring.jpa.show-sql}")
     private boolean showSql;
 
-    private int maximumPoolSize =10;
-
-    @Bean(name = "orderResultDatasource")
-    public HikariDataSource orderResultDataSource() {
+    @Bean
+    @Primary
+    public DataSource dataSource() {
         HikariConfig hikariConfig = new HikariConfig();
         hikariConfig.setJdbcUrl(url);
         hikariConfig.setUsername(username);
         hikariConfig.setPassword(password);
         hikariConfig.setMaximumPoolSize(maximumPoolSize);
-        hikariConfig.setConnectionTimeout(15000L);
-        hikariConfig.setPoolName("orderResultPool");
-        //orderEventQuiz만을 위한 설정 0.1초마다 확인
+        hikariConfig.setLeakDetectionThreshold(leakDetectionThreshold);
+        hikariConfig.setPoolName("orderPool");
         return new HikariDataSource(hikariConfig);
     }
-    @Bean(name ="orderResultEntityManager")
-    public LocalContainerEntityManagerFactoryBean orderResultEntityManager(EntityManagerFactoryBuilder builder)  {
+
+    @Bean
+    @Primary
+    public LocalContainerEntityManagerFactoryBean entityManagerFactory(
+            EntityManagerFactoryBuilder builder
+    ) {
         Map<String,Object> properties = new HashMap<>();
         properties.put("dialect", "org.hibernate.dialect.MySQL8InnoDBDialect");
         properties.put("hibernate.show_sql", showSql);
         properties.put("hibernate.format_sql", true);
-//        properties.put("hibernate.ddl-auto", "create");
+//        properties.put("hibernate.ddl-auto", "update");
         properties.put("open_in_view", "false");
         properties.put("hibernate.hbm2ddl.auto", "update");
         properties.put("hibernate.physical_naming_strategy" , "org.hibernate.boot.model.naming.CamelCaseToUnderscoresNamingStrategy");
         return builder
-                .dataSource(orderResultDataSource())
-                .packages("com.watermelon.server.orderResult")
+                .dataSource(dataSource())
+                .packages("com.watermelon.server.order")
                 .properties(properties)
                 .build();
     }
-    @Bean(name = "orderResultTransactionManager")
-    public JpaTransactionManager orderResultTransactionManager(
-            @Qualifier("orderResultEntityManager") LocalContainerEntityManagerFactoryBean entityManagerFactory
-    ) {
+
+    @Bean
+    @Primary
+    public JpaTransactionManager transactionManager(LocalContainerEntityManagerFactoryBean entityManagerFactory) {
         return new JpaTransactionManager(entityManagerFactory.getObject());
     }
 }
