+++ order/src/main/java/com/watermelon/server/common/dataSourceConfig/OrderResultDataSourceConfig.java
@@ -1,34 +1,31 @@
-package com.watermelon.server.common.config;
+package com.watermelon.server.common.dataSourceConfig;
+
 
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.orm.jpa.EntityManagerFactoryBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
-import javax.sql.DataSource;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Properties;
 
 @EnableTransactionManagement
 @EnableJpaRepositories(
-        entityManagerFactoryRef = "entityManagerFactory",
-        transactionManagerRef = "transactionManager",
-        basePackages = {"com.watermelon.server.order.repository"}
+        entityManagerFactoryRef = "orderResultEntityManager",
+        transactionManagerRef = "orderResultTransactionManager",
+        basePackages = {"com.watermelon.server.orderResult.repository"}
 )
 @Configuration
 @Profile("!local")
-public class DatasourceConfig {
-
+public class OrderResultDataSourceConfig {
     @Value("${spring.datasource.url}")
     private String url;
 
@@ -38,51 +35,43 @@ public class DatasourceConfig {
     @Value("${spring.datasource.password}")
     private String password;
 
-    @Value("${spring.datasource.hikari.maximum-pool-size}")
-    private int maximumPoolSize;
-
-    @Value("${spring.datasource.hikari.leak-detection-threshold}")
-    private long leakDetectionThreshold;
-
     @Value("${spring.jpa.show-sql}")
     private boolean showSql;
 
-    @Bean
-    @Primary
-    public DataSource dataSource() {
+    private int maximumPoolSize =10;
+
+    @Bean(name = "orderResultDatasource")
+    public HikariDataSource orderResultDataSource() {
         HikariConfig hikariConfig = new HikariConfig();
         hikariConfig.setJdbcUrl(url);
         hikariConfig.setUsername(username);
         hikariConfig.setPassword(password);
         hikariConfig.setMaximumPoolSize(maximumPoolSize);
-        hikariConfig.setLeakDetectionThreshold(leakDetectionThreshold);
-        hikariConfig.setPoolName("orderPool");
+        hikariConfig.setConnectionTimeout(15000L);
+        hikariConfig.setPoolName("orderResultPool");
+        //orderEventQuiz만을 위한 설정 0.1초마다 확인
         return new HikariDataSource(hikariConfig);
     }
-
-    @Bean
-    @Primary
-    public LocalContainerEntityManagerFactoryBean entityManagerFactory(
-            EntityManagerFactoryBuilder builder
-    ) {
+    @Bean(name ="orderResultEntityManager")
+    public LocalContainerEntityManagerFactoryBean orderResultEntityManager(EntityManagerFactoryBuilder builder)  {
         Map<String,Object> properties = new HashMap<>();
         properties.put("dialect", "org.hibernate.dialect.MySQL8InnoDBDialect");
         properties.put("hibernate.show_sql", showSql);
         properties.put("hibernate.format_sql", true);
-//        properties.put("hibernate.ddl-auto", "update");
+//        properties.put("hibernate.ddl-auto", "create");
         properties.put("open_in_view", "false");
         properties.put("hibernate.hbm2ddl.auto", "update");
         properties.put("hibernate.physical_naming_strategy" , "org.hibernate.boot.model.naming.CamelCaseToUnderscoresNamingStrategy");
         return builder
-                .dataSource(dataSource())
-                .packages("com.watermelon.server.order")
+                .dataSource(orderResultDataSource())
+                .packages("com.watermelon.server.orderResult")
                 .properties(properties)
                 .build();
     }
-
-    @Bean
-    @Primary
-    public JpaTransactionManager transactionManager(LocalContainerEntityManagerFactoryBean entityManagerFactory) {
+    @Bean(name = "orderResultTransactionManager")
+    public JpaTransactionManager orderResultTransactionManager(
+            @Qualifier("orderResultEntityManager") LocalContainerEntityManagerFactoryBean entityManagerFactory
+    ) {
         return new JpaTransactionManager(entityManagerFactory.getObject());
     }
 }
