+++ order/src/main/java/com/watermelon/server/order/service/OrderResultSaveService.java
@@ -18,7 +18,7 @@ public class OrderResultSaveService {
     private final OrderEventFromServerMemoryService orderEventFromServerMemoryService;
 
 
-    @Transactional(transactionManager = "orderResultTransactionManager")
+
     public boolean isOrderApplyNotFullThenSaveConnectionOpen(String applyToken,int applyCountIndex) throws CannotCreateTransactionException {
         /**
          * 먼저 락을 걸고 ApplyCount의 숫자를 올리는 메소드가 성공하였다면
