+++ order/src/test/java/com/watermelon/server/order/service/OrderEventFromServerMemoryServiceTest.java
@@ -4,14 +4,12 @@ import com.watermelon.server.order.domain.OrderEvent;
 import com.watermelon.server.order.dto.request.RequestOrderEventDto;
 import com.watermelon.server.order.dto.request.RequestOrderRewardDto;
 import com.watermelon.server.order.dto.request.RequestQuizDto;
-import com.watermelon.server.orderResult.repository.OrderApplyCountRepository;
+import com.watermelon.server.orderApplyCount.repository.OrderApplyCountRepository;
 import com.watermelon.server.order.repository.OrderEventRepository;
-import com.watermelon.server.orderResult.repository.OrderResultRepository;
-import com.watermelon.server.orderResult.domain.OrderApplyCount;
-import com.watermelon.server.orderResult.service.OrderApplyCountService;
+import com.watermelon.server.orderApplyCount.domain.OrderApplyCount;
+import com.watermelon.server.orderApplyCount.service.OrderApplyCountService;
 import com.watermelon.server.orderResult.service.OrderEventFromServerMemoryService;
 import com.watermelon.server.orderResult.service.IndexLoadBalanceService;
-import com.watermelon.server.orderResult.service.OrderApplyCountLockService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -19,7 +17,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.List;
