+++ order/src/test/java/com/watermelon/server/order/total/OrderEventTotalTest.java
@@ -7,9 +7,9 @@ import com.watermelon.server.order.domain.OrderEvent;
 import com.watermelon.server.order.domain.OrderEventStatus;
 import com.watermelon.server.order.domain.Quiz;
 import com.watermelon.server.order.dto.request.*;
-import com.watermelon.server.orderResult.repository.OrderApplyCountRepository;
+import com.watermelon.server.orderApplyCount.repository.OrderApplyCountRepository;
 import com.watermelon.server.order.repository.OrderEventRepository;
-import com.watermelon.server.orderResult.domain.OrderApplyCount;
+import com.watermelon.server.orderApplyCount.domain.OrderApplyCount;
 import com.watermelon.server.orderResult.service.OrderEventFromServerMemoryService;
 import com.watermelon.server.token.ApplyTokenProvider;
 import com.watermelon.server.token.JwtPayload;
